+++ src/lib.rs
@@ -157,13 +157,17 @@
 use aws_config::meta::region::RegionProviderChain;
 use aws_sdk_s3::{Client, Endpoint, Region};
 
+#[macro_use]
+extern crate derive_builder;
+
 use futures::stream::StreamExt;
 use http::Uri;
 use tokio::sync::mpsc;
 
 pub use near_indexer_primitives;
 
-pub use types::LakeConfig;
+pub use aws_types::Credentials;
+pub use types::{LakeConfig, LakeConfigBuilder};
 
 mod s3_fetchers;
 pub(crate) mod types;
@@ -191,32 +195,24 @@ pub(crate) const LAKE_FRAMEWORK: &str = "near_lake_framework";
 /// ```
 pub fn streamer(config: LakeConfig) -> mpsc::Receiver<near_indexer_primitives::StreamerMessage> {
     let (sender, receiver) = mpsc::channel(100);
-    tokio::spawn(start(
-        sender,
-        config.s3_endpoint,
-        config.s3_bucket_name,
-        config.s3_region_name,
-        config.start_block_height,
-    ));
+    tokio::spawn(start(sender, config));
     receiver
 }
 
 async fn start(
     streamer_message_sink: mpsc::Sender<near_indexer_primitives::StreamerMessage>,
-    s3_endpoint: Option<String>,
-    s3_bucket_name: String,
-    s3_region_name: String,
-    index_from_block_height: types::BlockHeight,
+    config: LakeConfig,
 ) {
     // instantiate AWS S3 Client
-    let region_provider = RegionProviderChain::first_try(Some(s3_region_name).map(Region::new))
-        .or_default_provider()
-        .or_else(Region::new("eu-central-1"));
+    let region_provider =
+        RegionProviderChain::first_try(Some(config.s3_region_name).map(Region::new))
+            .or_default_provider()
+            .or_else(Region::new("eu-central-1"));
     let shared_config = aws_config::from_env().region(region_provider).load().await;
     let mut s3_conf = aws_sdk_s3::config::Builder::from(&shared_config);
     // Owerride S3 endpoint in case you want to use custom solution
     // like Minio or Localstack as a S3 compatible storage
-    if let Some(endpoint) = s3_endpoint {
+    if let Some(endpoint) = config.s3_endpoint {
         s3_conf = s3_conf.endpoint_resolver(Endpoint::immutable(endpoint.parse::<Uri>().unwrap()));
         tracing::info!(
             target: LAKE_FRAMEWORK,
@@ -224,15 +220,21 @@ async fn start(
             endpoint
         );
     }
+
+    if let Some(credentials) = config.s3_credentials {
+        s3_conf = s3_conf.credentials_provider(credentials);
+    }
+
     let s3_client = Client::from_conf(s3_conf.build());
 
-    let mut start_from_block_height = index_from_block_height;
+    let mut start_from_block_height = config.start_block_height;
     let mut last_processed_block_hash: Option<near_indexer_primitives::CryptoHash> = None;
 
     // Continuously get the list of block data from S3 and send them to the `streamer_message_sink`
     loop {
         if let Ok(block_heights_prefixes) =
-            s3_fetchers::list_blocks(&s3_client, &s3_bucket_name, start_from_block_height).await
+            s3_fetchers::list_blocks(&s3_client, &config.s3_bucket_name, start_from_block_height)
+                .await
         {
             if block_heights_prefixes.is_empty() {
                 tracing::debug!(
@@ -253,7 +255,7 @@ async fn start(
                     .map(|block_height| {
                         s3_fetchers::fetch_streamer_message(
                             &s3_client,
-                            &s3_bucket_name,
+                            &config.s3_bucket_name,
                             *block_height,
                         )
                     })
@@ -286,7 +288,7 @@ async fn start(
             tracing::error!(
                 target: LAKE_FRAMEWORK,
                 "Failed to list objects from bucket {}. Retrying...",
-                &s3_bucket_name
+                &config.s3_bucket_name
             );
         }
     }