+++ src/types.rs
@@ -20,8 +20,8 @@ pub struct LakeConfig {
     ///
     /// In case you want to run your own [near-lake](https://github.com/near/near-lake) instance and store data in some S3 compatible storage ([Minio](https://min.io/) or [Localstack](https://localstack.cloud/) as example)
     ///
-    /// You can owerride default S3 API endpoint by using `s3_endpoint` option or [LakeConfigBuilder::with_custom_endpoint]"]
-    #[builder(setter(custom, strip_option, name = "with_custom_endpoint"), default)]
+    /// You can owerride default S3 API endpoint by using `s3_endpoint` option or [LakeConfigBuilder::custom_endpoint]"]
+    #[builder(setter(custom, strip_option, name = "custom_endpoint"), default)]
     pub s3_endpoint: Option<String>,
     /// AWS S3 Bucket name
     #[builder(setter(into))]
@@ -31,15 +31,12 @@ pub struct LakeConfig {
     pub s3_region_name: String,
     /// Defines the block height to start indexing from
     pub start_block_height: u64,
-    #[builder(
-        setter(custom, strip_option, name = "with_custom_credentials"),
-        default
-    )]
+    #[builder(setter(custom, strip_option, name = "custom_credentials"), default)]
     /// AWS S3 Credentials
     /// By default NEAR Lake Framework will use the AWS credentials stored in the file `~/.aws/credentials`.
     /// However, sometimes you might want to pass the credentials through the CLI arguments. You can set [LakeConfig::s3_credentials] instead of using the file.
     ///
-    /// See [LakeConfigBuilder::with_custom_credentials]
+    /// See [LakeConfigBuilder::custom_credentials]
     pub s3_credentials: Option<aws_types::Credentials>,
 }
 
@@ -48,19 +45,19 @@ impl LakeConfigBuilder {
     ///
     /// In case you want to run your own [near-lake](https://github.com/near/near-lake) instance and store data in some S3 compatible storage ([Minio](https://min.io/) or [Localstack](https://localstack.cloud/) as example)
     ///
-    /// You can owerride default S3 API endpoint by using `with_custom_endpoint` method
+    /// You can owerride default S3 API endpoint by using `custom_endpoint` method
     /// ```
     /// use near_lake_framework::LakeConfigBuilder;
     /// # async fn main() {
     /// let config = LakeConfigBuilder::default()
     ///     .s3_bucket_name("near-lake-data-testnet")
     ///     .start_block_height(82422587)
-    ///     .with_custom_endpoint("http://0.0.0.0:9000")
+    ///     .custom_endpoint("http://0.0.0.0:9000")
     ///     .build()
     ///     .expect("Failed to build LakeConfig");
     /// # }
     /// ```
-    pub fn with_custom_endpoint(&mut self, endpoint: impl Into<String>) -> &mut Self {
+    pub fn custom_endpoint(&mut self, endpoint: impl Into<String>) -> &mut Self {
         self.s3_endpoint = Some(Some(endpoint.into()));
         self
     }
@@ -74,7 +71,7 @@ impl LakeConfigBuilder {
     /// let config = LakeConfigBuilder::default()
     ///     .s3_bucket_name("near-lake-data-testnet")
     ///     .start_block_height(82422587)
-    ///     .with_custom_credentials(
+    ///     .custom_credentials(
     ///         "AKIAIOSFODNN7EXAMPLE",
     ///         "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
     ///     )
@@ -82,7 +79,7 @@ impl LakeConfigBuilder {
     ///     .expect("Failed to build LakeConfig");
     /// # }
     /// ```
-    pub fn with_custom_credentials(
+    pub fn custom_credentials(
         &mut self,
         access_key_id: impl Into<String>,
         secret_access_key: impl Into<String>,