+++ CHANGELOG.md
@@ -21,8 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   use http::Uri;
   use near_lake_framework::LakeConfigBuilder;
 
-
-  let mut s3_conf = aws_sdk_s3::config::Builder::from(&shared_config);
+  let aws_config = aws_config::from_env().load().await;
+  let mut s3_conf = aws_sdk_s3::config::Builder::from(&aws_config);
   s3_conf = s3_conf
     .endpoint_resolver(
       Endpoint::immutable("http://0.0.0.0:9000".parse::<Uri>().unwrap()))